Share rosterOffersAChoice with the sheet's trigger surfaces (#2111)
Moves the #1560 predicate out of the FieldDesk page into
hooks/useRosterChoice so the rail's character card and the eight mobile
skins can ask the same question without importing a page module, and
carries the answer to the skins on FieldDeskHomeState.
